backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
PyMuPDF>=1.23.0
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
